+++ src/data_structs.hpp
@@ -1,7 +1,9 @@
 #include <cstdint>
 #include "comms/data/comms_data.hpp"
 #include "sensor_constants.hpp"
-#include "src/sensors/d200.hpp"
+
+#ifndef DATA_STRUCTS_HPP
+#define DATA_STRUCTS_HPP
 
 /// @brief Structure for the buff encoder sensor.
 struct BuffEncoderData : Comms::CommsData {
@@ -44,18 +46,6 @@ struct ICMSensorData : Comms::CommsData {
     float temperature;
 };
 
-/// @brief Structure for the Rev encoder sensor.
-struct RevSensorData : Comms::CommsData {
-    RevSensorData() : CommsData(Comms::TypeLabel::RevEncoderData, Comms::PhysicalMedium::Ethernet, Comms::Priority::Medium, sizeof(RevSensorData)) { }
-	/// Sensor ID.
-	uint8_t id;
-	/// Encoder ticks.
-	int ticks;
-	/// Angle in radians.
-	float radians;
-    
-};
-
 /// @brief Structure for the TOF (Time-of-Flight) sensor.
 struct TOFSensorData : Comms::CommsData {
     TOFSensorData() : CommsData(Comms::TypeLabel::TOFSensorData, Comms::PhysicalMedium::Ethernet, Comms::Priority::Medium, sizeof(TOFSensorData)) { }
@@ -64,6 +54,49 @@ struct TOFSensorData : Comms::CommsData {
     /// Latest distance measurement.
     uint16_t latest_distance;
 };
+
+/// @brief data for a LiDAR packet (SI units)
+struct LidarDataPacketSI {
+    /// @brief speed of lidar module (rad/s)
+    float lidar_speed = 0;
+  
+    /// @brief start angle of measurements (rad)
+    float start_angle = 0;
+  
+    /// @brief array of point measurements
+    struct {
+      /// @brief distance (m)
+      float distance;
+  
+      /// @brief intensity of measurement. units are ambiguous (not documented), but in general "the higher the intensity, the larger the signal strength value"
+      uint8_t intensity = 0;
+    } points[D200_POINTS_PER_PACKET];
+  
+    /// @brief end angle of measurements (rad)
+    float end_angle = 0;
+  
+    /// @brief timestamp of measurements, calibrated (s)
+    float timestamp = 0;
+  };
+  
+  /// @brief struct storing timestamp calibration results 
+  struct D200Calibration {
+    /// @brief how many packets are used for calibration
+    int max_calibration_packets = D200_MAX_CALIBRATION_PACKETS;
+  
+    /// @brief how many calibration packets have been received
+    int packets_recv = 0;
+  
+    /// @brief count of D200 timestamp wraps
+    int num_wraps = 0;
+  
+    /// @brief previous lidar timestamp
+    int prev_timestamp = -1;
+  
+    /// @brief sum of delta times for calibration packets
+    int timestamp_delta_sum = 0;
+  };
+  
 
 /// @brief Structure for the LiDAR sensor.
 struct LidarSensorData : Comms::CommsData {
@@ -125,3 +158,4 @@ struct DR16Data : Comms::CommsData {
     };
 };
 
+#endif // DATA_STRUCTS_HPP
